requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.3
mistune==3.0.2
PyYAML==6.0.1